+++ BasicData/BasicData.Web/UI_BasicData/EnergyDataManualInput/EnergyDataManualInputContrast.aspx.cs
@@ -11,14 +11,32 @@ namespace BasicData.Web.UI_BasicData.EnergyDataManualInput
 {
     public partial class EnergyDataManualInputContrast : WebStyleBaseForEnergy.webStyleBase
     {
+        private static char[] CRUD; 
         protected void Page_Load(object sender, EventArgs e)
         {
             base.InitComponts();
             ////////////////////调试用,自定义的数据授权
             //List<string> m_DataValidIdItems = new List<string>() { "C41B1F47-A48A-495F-A890-0AABB2F3BFF7                            ", "43F1EA8C-FF77-4BC5-BACB-531DC56A2512                            " };
             //AddDataValidIdGroup("ProductionOrganization", m_DataValidIdItems);
+#if DEBUG
+            ////////////////////调试用,自定义的数据授权
+            List<string> m_DataValidIdItems = new List<string>() { "zc_nxjc_byc" };
+            AddDataValidIdGroup("ProductionOrganization", m_DataValidIdItems);
+            mPageOpPermission = "0000";
+            
+#elif RELEASE
+#endif
+            CRUD = mPageOpPermission.ToArray();
         }
-
+        /// <summary>
+        /// 增删改查权限控制
+        /// </summary>
+        /// <returns></returns>
+        [WebMethod]
+        public static char[] AuthorityControl()
+        {
+            return mPageOpPermission.ToArray();
+        }
         [WebMethod]
         public static string GetEnergyDataManualInputContrastData(string variableName)
         {
@@ -30,36 +48,57 @@ namespace BasicData.Web.UI_BasicData.EnergyDataManualInput
         [WebMethod]
         public static string AddEnergyDataManualInputContrastData(string maddData)
         {
-            int result = BasicData.Service.EnergyDataManualInput.EnergyDataManualInputService.AddEnergyDataManualInputContrast(maddData);
+            if (CRUD[1] == '1')
+            {
+                int result = BasicData.Service.EnergyDataManualInput.EnergyDataManualInputService.AddEnergyDataManualInputContrast(maddData);
 
-            if (result == 1)
-                return "1";
-            else if (result == -2)
-                return "-2";
+                if (result == 1)
+                    return "1";
+                else if (result == -2)
+                    return "-2";
+                else
+                    return "-1";
+            }
             else
-                return "-1";
+            {
+                return "noright";
+            }
         }
 
         [WebMethod]
         public static string DeleteEnergyDataManualInputContrastData(string variableId)
         {
-            int result = BasicData.Service.EnergyDataManualInput.EnergyDataManualInputService.DeleteEnergyDataManualInputContrast(variableId);
+            if (CRUD[3] == '1')
+            {
+                int result = BasicData.Service.EnergyDataManualInput.EnergyDataManualInputService.DeleteEnergyDataManualInputContrast(variableId);
 
-            if (result == 1)
-                return "1";
+                if (result == 1)
+                    return "1";
+                else
+                    return "-1";
+            }
             else
-                return "-1";
+            {
+                return "noright";
+            }
         }
 
         [WebMethod]
         public static string EditEnergyDataManualInputContrastData(string editData)
         {
-            int result = BasicData.Service.EnergyDataManualInput.EnergyDataManualInputService.EditEnergyDataManualInputContrast(editData);
+            if (CRUD[2] == '1')
+            {
+                int result = BasicData.Service.EnergyDataManualInput.EnergyDataManualInputService.EditEnergyDataManualInputContrast(editData);
 
-            if (result == 1)
-                return "1";
+                if (result == 1)
+                    return "1";
+                else
+                    return "-1";
+            }
             else
-                return "-1";
+            {
+                return "noright";
+            }
         }
     }
 }
