+++ BasicData/BasicData.Web/UI_BasicData/EquipmentAccount/EquipmentAccountEdit.aspx.cs
@@ -12,6 +12,7 @@ namespace BasicData.Web.UI_BasicData.EquipmentAccount
 {
     public partial class EquipmentAccountEdit : WebStyleBaseForEnergy.webStyleBase
     {
+        
         protected void Page_Load(object sender, EventArgs e)
         {
             base.InitComponts();
@@ -21,11 +22,21 @@ namespace BasicData.Web.UI_BasicData.EquipmentAccount
 #if DEBUG
                 List<string> m_DataValidIdItems = new List<string>() { "zc_nxjc_byc_byf" };
                 AddDataValidIdGroup("ProductionOrganization", m_DataValidIdItems);
+                mPageOpPermission = "0000";
 #elif RELEASE
 #endif
                 this.OrganisationTree_ProductionLine.Organizations = GetDataValidIdGroup("ProductionOrganization");                 //向web用户控件传递数据授权参数
                 this.OrganisationTree_ProductionLine.PageName = "EquipmentAccountEdit.aspx";                                     //向web用户控件传递当前调用的页面名称
             }
+        }
+        /// <summary>
+        /// 增删改查权限控制
+        /// </summary>
+        /// <returns></returns>
+        [WebMethod]
+        public static char[] AuthorityControl()
+        {
+            return mPageOpPermission.ToArray();
         }
         /// <summary>
         /// 获取设备信息
@@ -67,9 +78,16 @@ namespace BasicData.Web.UI_BasicData.EquipmentAccount
         public static string SaveEquipmentInfo(string VariableId, string OrganizationID, string EquipmentName, int MonitorType,string PowerSupply, string VoltageGrade, 
             string RatedCT, string AmmeterCode, string ActualCT, int Power, string Unit,string Current,string PowerSupplyPosition, string Remarks)
         {
-            string message= EquipmentAccountService.SaveEquipment(VariableId, OrganizationID, EquipmentName, MonitorType, PowerSupply, VoltageGrade,RatedCT, AmmeterCode, ActualCT, 
-                Power, Unit,Current, PowerSupplyPosition, Remarks);
-            return message;
+            if (mPageOpPermission.ToArray()[1] == '1')
+            {
+                string message = EquipmentAccountService.SaveEquipment(VariableId, OrganizationID, EquipmentName, MonitorType, PowerSupply, VoltageGrade, RatedCT, AmmeterCode, ActualCT,
+                    Power, Unit, Current, PowerSupplyPosition, Remarks);
+                return message;
+            }
+            else
+            {
+                return "用户没有修改权限！";
+            }
         }
         /// <summary>
         /// 删除设备信息
@@ -80,7 +98,14 @@ namespace BasicData.Web.UI_BasicData.EquipmentAccount
         [WebMethod]
         public static string RemoveEquipmentInfo(string variableId, string organizationId)
         {
-            return EquipmentAccountService.RemoveEquipment(variableId, organizationId);
+            if (mPageOpPermission.ToArray()[3] == '1')
+            {
+                return EquipmentAccountService.RemoveEquipment(variableId, organizationId);
+            }
+            else
+            {
+                return "用户没有删除权限！";
+            }
         }
         /// <summary>
         /// 编辑设备
@@ -92,8 +117,15 @@ namespace BasicData.Web.UI_BasicData.EquipmentAccount
         public static string EditEquipmentInfo(string variableId_old, string organizationId_old, string VariableId, string OrganizationID, string EquipmentName, int MonitorType, string PowerSupply, string VoltageGrade,
             string RatedCT, string AmmeterCode, string ActualCT, int Power, string Unit, string Current,string PowerSupplyPosition, string Remarks)
         {
-            return EquipmentAccountService.UpdateEquipment(variableId_old, organizationId_old, VariableId, OrganizationID, EquipmentName, MonitorType, PowerSupply,
-                VoltageGrade, RatedCT, AmmeterCode, ActualCT, Power, Unit, Current,PowerSupplyPosition, Remarks);
+            if (mPageOpPermission.ToArray()[2] == '1')
+            {
+                return EquipmentAccountService.UpdateEquipment(variableId_old, organizationId_old, VariableId, OrganizationID, EquipmentName, MonitorType, PowerSupply,
+                    VoltageGrade, RatedCT, AmmeterCode, ActualCT, Power, Unit, Current, PowerSupplyPosition, Remarks);
+            }
+            else
+            {
+                return "用户没有编辑权限！";
+            }
         }
     }
 }
